+++ patchworker.py
@@ -37,9 +37,7 @@ class Resume:
         self.location = XmlHelper.findtext(element, "contact/location")
         self.linkedin = XmlHelper.findtext(element, "contact/linkedin")
         self.github = XmlHelper.findtext(element, "contact/github")
-        self.experience_sections = [ExperienceSection(exps_el) for exps_el in element.findall("experiences")]
-        self.education_section = EducationSection(element.findall("education")[0])
-        self.skill_section = SkillSection(element.findall("skills")[0])
+        self.sections = [Section(sec_el) for sec_el in element.findall("section")]
 
     def to_latex(self):
         latex = "\\documentclass{patchworker}\n"
@@ -54,23 +52,38 @@ github={{{escape_latex(self.github)}}}
 }}
 """
         latex += "\n\\begin{document}\n\n"
-        latex += self.skill_section.to_latex()
-        for exp_sec in self.experience_sections:
-            latex += exp_sec.to_latex()
-        latex += self.education_section.to_latex()
+        for sec in self.sections:
+            latex += sec.to_latex()
         latex += "\\end{document}\n"
         return latex
 
-class ExperienceSection:
+class Section:
+    def __str__(self):
+        return f"Section({self.type=}, {self.heading=}, {self.children.count()})"
+
     def __init__(self, element):
         self.type = element.attrib.get("type")
-        self.heading = element.attrib.get("heading", self.type.capitalize() + " Experience")
-        self.experiences = [Experience(exp_el) for exp_el in element.findall("experience")]
+        self.heading = element.attrib.get("heading", "Section")
+        self.children = []
+        for child in element:
+            obj = None
+            if child.tag == "experience":
+                obj = Experience(child)
+            elif child.tag == "degree":
+                obj = Degree(child)
+            elif child.tag == "skills":
+                obj = SkillSection(child)
+            elif child.tag == "description":
+                obj = Description(child)
+            else:
+                raise Exception(f"Don't know how to handle `{child.tag}' child")
+            self.children.append(obj)
+        #self.children = [Experience(exp_el) for exp_el in element.findall("experience")]
 
     def to_latex(self):
         latex = f"\\section*{{{escape_latex(self.heading)}}}\n"
-        for exp in self.experiences:
-            latex += exp.to_latex() + "\n"
+        for child in self.children:
+            latex += child.to_latex() + "\n"
         return latex
 
 class Experience:
@@ -101,17 +114,6 @@ class Experience:
         latex += "\\end{itemize}\n"
         return latex
 
-class EducationSection:
-    def __init__(self, element):
-        self.heading = element.attrib.get("heading", "Education")
-        self.degrees = [Degree(deg_el) for deg_el in element.findall("degree")]
-
-    def to_latex(self):
-        latex = f"\\section*{{{escape_latex(self.heading)}}}\n"
-        for deg in self.degrees:
-            latex += deg.to_latex()
-        return latex
-
 class Degree:
     def __init__(self, element):
         self.date = XmlHelper.findtext(element, "date")
@@ -131,11 +133,10 @@ class Degree:
 
 class SkillSection:
     def __init__(self, element):
-        self.heading = element.attrib.get("heading", "Skills")
         self.skills = [Skill(skill_el) for skill_el in element.findall("skill")]
 
     def to_latex(self):
-        latex = f"\\begin{{skills}}{{{escape_latex(self.heading)}}}\n"
+        latex = f"\\begin{{skills}}\n"
         for skill in self.skills:
             latex += f"\\item {skill.to_latex()}\n"
         latex += "\n\\end{skills}\n"
@@ -147,6 +148,13 @@ class Skill:
 
     def to_latex(self):
         return escape_latex(self.name)
+
+class Description:
+    def __init__(self, element):
+        self.text = element.text
+
+    def to_latex(self):
+        return escape_latex(self.text)
 
 def main():
     parser = argparse.ArgumentParser(description="Generate LaTeX resume from XML")